requirements.txt
numpy==1.26.4
pandas==2.3.0
scipy==1.16.1
scikit-learn==1.7.0
scikit-image==0.25.2
matplotlib==3.10.3
opencv-python-headless==4.11.0.86
ipykernel==6.29.5
ipycanvas==0.14.1
ipympl==0.9.7
pillow==11.2.1
pydantic==2.11.7
pytest==9.0.2
mypy==1.19.1
scipy-stubs
ultralytics
pytorch-msssim==1.0.0
wandb
dotenv
